+++ app/ui/components/project-views/branches/branch-new.jsx
@@ -42,8 +42,7 @@ class CreateBranch extends Component {
 
     // Initialize state props
     this.state = {
-      branches: null,
-      tags: null,
+      branches: [],
       name: '',
       id: '',
       source: 'master',
@@ -130,7 +129,7 @@ class CreateBranch extends Component {
           this.setState({ error: err.responseText });
         },
         401: () => {
-          this.setState({ branches: null, tags: null });
+          this.setState({ branches: [] });
 
           // Refresh when session expires
           window.location.reload();
@@ -151,29 +150,18 @@ class CreateBranch extends Component {
     let idInvalid = false;
     let customInvalid = false;
 
-    const branchOptions = [];
-    const tagOptions = [];
+    const branches = [];
+    const tags = [];
 
-    if (this.state.branches) {
-      this.state.branches.forEach((branch) => {
-        if (!branch.tag) {
-          branchOptions.push(
-            <option className='branch-opts'
-                    value={branch.id}>
-              {(branch.name.length > 0) ? branch.name : branch.id}
-            </option>
-          );
-        }
-        else {
-          tagOptions.push(
-            <option className='branch-opts'
-                    value={branch.id}>
-              {(branch.name.length > 0) ? branch.name : branch.id}
-            </option>
-          );
-        }
-      });
-    }
+    this.state.branches.forEach((branch, idx) => {
+      const option = (<option key={`opt-${idx}`} className='branch-opts' value={branch.id}>
+                        {(branch.name.length > 0) ? branch.name : branch.id}
+                      </option>);
+
+      const options = (branch.tag) ? tags : branches;
+
+      options.push(option);
+    });
 
     // Verify if id is valid
     const { id } = this.state;
@@ -220,9 +208,9 @@ class CreateBranch extends Component {
                      value={this.state.source || ''}
                      onChange={this.handleChange}>
                 <option disabled={true}>Branches</option>
-                {branchOptions}
+                {branches}
                 <option disabled={true}>Tags</option>
-                {tagOptions}
+                {tags}
               </Input>
             </FormGroup>
             {/* Create an input for project id */}